requirements.txt
Flask
Flask-Bootstrap4
Flask-WTF
jinja2
WTForms
Werkzeug
hdbcli
pandas
PyYAML
fitparse
